Making configureStore compatible with Webpack 2
