--- src/store/configureStore.dev.js
@@ -1,36 +1,0 @@
-// This file merely configures the store for hot reloading.
-// This boilerplate file is likely to be the same for each project that uses Redux.
-// With Redux, the actual stores are in /reducers.
-
-import { createStore, applyMiddleware, compose } from "redux";
-import reduxImmutableStateInvariant from "redux-immutable-state-invariant";
-import createSagaMiddleware from "redux-saga";
-import root from "../middleware/sagas";
-import rootReducer from "../reducers";
-
-export default function configureStore(initialState) {
-  const middlewares = [
-    // Add other middleware on this line...
-    createSagaMiddleware(),
-    reduxImmutableStateInvariant(),
-  ];
-
-  // add support for Redux dev tools
-  const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose; // eslint-disable-line no-underscore-dangle
-  const store = createStore(rootReducer, initialState, composeEnhancers(
-    applyMiddleware(...middlewares)
-    )
-  );
-
-  middlewares[0].run(root);
-
-  if (module.hot) {
-    // Enable Webpack hot module replacement for reducers
-    module.hot.accept("../reducers", () => {
-      const nextReducer = require("../reducers").default;
-      store.replaceReducer(nextReducer);
-    });
-  }
-
-  return store;
-}
