+++ webpack.config.analyzer.js
@@ -1,5 +1,5 @@
-import {BundleAnalyzerPlugin} from 'webpack-bundle-analyzer';
-import config from './webpack.config.dev';
+import {BundleAnalyzerPlugin} from "webpack-bundle-analyzer";
+import config from "./webpack.config.dev";
 
 config.plugins.push(new BundleAnalyzerPlugin());
